Update 'GeomRangeFrame' to handle NAs
'geom_rangeframe' was not giving a warning when 'na.rm = FALSE', as was
documented. In addition, the range lines were not being drawn at all when there were
NAs in the data.

The handling of the required aesthetics has been
updated to depend on the value of 'sides', and now the warning is triggered
when 'na.rm = FALSE' and the range lines are still drawn.